Move derange and (isValid -> validateMatches) to derange.ts
